requirements.txt
click
rich
gitpython
openai
python-dotenv
